remap-vec.patch This patch changes how we avoid getting killed by an interrupt if the vectors at address zero and the u-boot we're running don't agree.
Instead of copying the vectors to SteppingStone, which introduced the very
ugly need to wait until AUX was released, and didn't work for NAND, we copy
the vectors to the beginning of our DRAM and map the address range 0x0...
to 0x3...

This obsoletes nor-irqvec.patch

We could probably also use the same logic to replace
uboot-s3c2410-norelocate_irqvec_cpy.patch
but then, never change a running system ;-)

Note: this is still experimental. Please give it a solid beating.

- Werner