uboot-force-backlight-on.patch
"Special feature" of pcf50633 -- a documented feature though --
means that you may need to perform a little sequence of actions to
get the backlight to come back on.

This patch does the sequence in U-Boot so that we no longer have
the sticky "off" state for the backlight if you reset while the
backlight was down in Linux.

Signed-off-by: Andy Green <andy@openmoko.com>