deadbat-defer-glamo.patch
This patch defers initialization of the Glamo core, and in particular
enabling the clocks, until we're sure to have good power. This reduces
the power budget we need in early boot by about 15mA.

This patch also cleans up the remains of related code that has already
been moved to a later boot phase.

Signed-off-by: Werner Almesberger <werner@openmoko.org>