deadbat-cpu-idle.patch
This patch makes the CPU enter IDLE mode while waiting in wait_for_power.

To wake up from IDLE mode, we need an interrupt. Note that this interrupt
only needs to be signaled to the CPU, but the CPU does not have to execute
it. However, it has to clear the interrupt or, upon returning from the
interrupt handler, the system will immediately be interrupted by the same
interrupt.

We have the following sources of external events:

- battery insertion / battery voltage rises above Vth(batok)
- USB insertion (giving us 100mA)
- adapter insertion (giving us 1A)
- USB host signals availability of 500mA
- 1Hz tick

USB signaling is already handled by an interrupt handler, so we just use
what's there. USB insertion and the 1Hz tick cause PMU interrupts, so we
need to enable the interrupt path from the PMU. Adapter insertion is
handled as part of USB insertion.

We don't wake up from IDLE on battery insertion or when the battery
voltage rises above the threshold. Instead, we just poll for it on each
1Hz tick.

Signed-off-by: Werner Almesberger <werner@openmoko.org>