deadbat-cpu-200.patch
Raise the CPU clock (FCLK) from 200MHz to 400MHz only after we've established
that we have a good power source.

That way, the little bit of processing that are done in wait_for_power cause
less spikes in current consumption, which decreases the risk of running out
of power while waiting for USB signaling or similar.

We also save about 7mA of static consumption.

We need to do this awkward two-step process, once in board_init and then
again in board_late_init, since various clocks that depend on PCLK are
calculated after board_init (and start.S uses yet another set of different
settings).

Signed-off-by: Werner Almesberger <werner@openmoko.org>