PATCH-fix-data-abort-from-sd-ombug799.patch.eml
This patch has been in the OE repository for a while, but was not in the
u-boot git repository. It is a workaround that allows u-boot to access
ext2/ext3 filesystems, removing the need for a dedicated FAT partition
to hold the kernel when booting from SD. I have only tested it on GTA01.

Signed-off-by: Mike Montour <mail@mmontour.net>